ai/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# ── SCRFD + ArcFace (primary detection & recognition) ───────────────
insightface>=0.7.3
# Pick one runtime:
onnxruntime>=1.16.0          ; platform_system != "Linux" or python_version < "3.12"
# onnxruntime-gpu>=1.16.0    # uncomment on GPU boxes; remove the CPU line above

# ── MTCNN + FaceNet (ensemble) ──────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
facenet-pytorch>=2.5.3

# ── Database (Supabase / PostgreSQL) ────────────────────────────────
psycopg2-binary>=2.9.0
pgvector>=0.2.5

# ── Optional: GAN-based face restoration for low-light enhancement ──
# Uncomment to swap the CLAHE baseline for a real GAN. Both are heavy.
# gfpgan>=1.3.8
# realesrgan>=0.3.0
# basicsr>=1.4.2
